Move Constants to com.example.examplemod package.
