+++ Common/src/main/java/com/example/examplemod/Constants.java
@@ -1,4 +1,4 @@
-package com.blamejared.stuff;
+package com.example.examplemod;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
